backend/requirements.txt
Django==6.0.1
djangorestframework
pandas
django-cors-headers
gunicorn